+++ frontend/vue.config.js
@@ -66,10 +66,6 @@ module.exports = {
         target: process.env.BASE_URL || 'http://localhost:5000/',
         changeOrigin: true
       },
-      '/saml': {
-        target: process.env.BASE_URL || 'http://localhost:5000/',
-        changeOrigin: true
-      },
       '/version': {
         target: process.env.BASE_URL || 'http://localhost:5000/',
         changeOrigin: true