+++ NEWS.rst
@@ -5,6 +5,7 @@ New features
 ------------
 * #42774: Add English translation. The user can now choose Danish or English screen texts by clicking the flags in the nav bar
 * #42869: Added delta-api parameter "changed_since" to /api/v1/ endpoints
+* #43544: Keycloak authentication. Keycloak container, DB and config added
 
 Bug fixes
 ---------